+++ src/library/DIPS.Mobile.UI/Components/ListItems/ListItem.cs
@@ -256,4 +256,9 @@ public partial class ListItem : ContentView
     private void SetTouchIsEnabled() => Touch.SetIsEnabled(Border, IsEnabled && Command is not null);
 
     private void BindToOptions(ListItemOptions? options) => options?.Bind(this);
+
+    private void AddContextMenu()
+    {
+        BindToOptions(ContextMenuOptions);
+    }
 }