+++ src/library/DIPS.Mobile.UI/Components/ListItems/Options/ContextMenu/ContextMenuOptions.cs
@@ -1,0 +1,12 @@
+using DIPS.Mobile.UI.Components.ContextMenus;
+
+namespace DIPS.Mobile.UI.Components.ListItems.Options.ContextMenu;
+
+public partial class ContextMenuOptions : ListItemOptions
+{
+    public override void DoBind(ListItem listItem)
+    {
+        ContextMenuEffect.SetMode(listItem.Border, Mode);
+        ContextMenuEffect.SetMenu(listItem.Border, Menu);
+    }
+}